requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
dataclasses-json==0.5.7
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
Jinja2==3.1.2
langchain==0.0.71
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.24.1
openai==0.26.1
packaging==23.0
pydantic==1.10.4
PyYAML==6.0
redis==4.4.2
requests==2.28.2
six==1.16.0
spotipy==2.22.0
SQLAlchemy==1.4.46
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
